переместил главную страничку в папку с хендлерами а индекс стал подобием роутера
