+++ routes/categories.php
@@ -1,0 +1,1 @@
+<h1>CATEGORIES</h1>
