+++ routes/test.php
@@ -1,0 +1,1 @@
+<h1>TEST</h1>
